requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.7
django-cors-headers==4.9.0
django-restframework==0.0.1
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.11.0


